Add user_tokens context
