+++ lib/elixir_google_scraper/account/user_tokens.ex
@@ -1,0 +1,120 @@
+defmodule ElixirGoogleScraper.Account.UserTokens do
+  import Ecto.Query, warn: false
+  alias ElixirGoogleScraper.Account.Schemas.UserToken
+
+  @hash_algorithm :sha256
+  @rand_size 32
+
+  # It is very important to keep the reset password token expiry short,
+  # since someone with access to the email may take over the account.
+  @reset_password_validity_in_days 1
+  @confirm_validity_in_days 7
+  @change_email_validity_in_days 7
+  @session_validity_in_days 60
+
+  def build_session_token(user) do
+    token = :crypto.strong_rand_bytes(@rand_size)
+
+    {token,
+     %UserToken{
+       token: token,
+       context: "session",
+       user_id: user.id
+     }}
+  end
+
+  def verify_session_token_query(token) do
+    query =
+      from(token in token_and_context_query(token, "session"),
+        join: user in assoc(token, :user),
+        where: token.inserted_at > ago(@session_validity_in_days, "day"),
+        select: user
+      )
+
+    {:ok, query}
+  end
+
+  @spec build_email_token(atom | %{:email => any, :id => any, optional(any) => any}, any) ::
+          {binary,
+           %ElixirGoogleScraper.Account.Schemas.UserToken{
+             __meta__: Ecto.Schema.Metadata.t(),
+             context: any,
+             id: nil,
+             inserted_at: nil,
+             sent_to: any,
+             token: binary,
+             user: Ecto.Association.NotLoaded.t(),
+             user_id: any
+           }}
+
+  def build_email_token(user, context) do
+    build_hashed_token(user, context, user.email)
+  end
+
+  defp build_hashed_token(user, context, sent_to) do
+    token = :crypto.strong_rand_bytes(@rand_size)
+    hashed_token = :crypto.hash(@hash_algorithm, token)
+
+    {Base.url_encode64(token, padding: false),
+     %UserToken{
+       token: hashed_token,
+       context: context,
+       sent_to: sent_to,
+       user_id: user.id
+     }}
+  end
+
+  def verify_email_token_query(token, context) do
+    case Base.url_decode64(token, padding: false) do
+      {:ok, decoded_token} ->
+        hashed_token = :crypto.hash(@hash_algorithm, decoded_token)
+        days = days_for_context(context)
+
+        query =
+          from(token in token_and_context_query(hashed_token, context),
+            join: user in assoc(token, :user),
+            where: token.inserted_at > ago(^days, "day") and token.sent_to == user.email,
+            select: user
+          )
+
+        {:ok, query}
+
+      :error ->
+        :error
+    end
+  end
+
+  defp days_for_context("confirm"), do: @confirm_validity_in_days
+  defp days_for_context("reset_password"), do: @reset_password_validity_in_days
+
+  def verify_change_email_token_query(token, context) do
+    case Base.url_decode64(token, padding: false) do
+      {:ok, decoded_token} ->
+        hashed_token = :crypto.hash(@hash_algorithm, decoded_token)
+
+        query =
+          from(token in token_and_context_query(hashed_token, context),
+            where: token.inserted_at > ago(@change_email_validity_in_days, "day")
+          )
+
+        {:ok, query}
+
+      :error ->
+        :error
+    end
+  end
+
+  def token_and_context_query(token, context) do
+    from(UserToken, where: [token: ^token, context: ^context])
+  end
+
+  def user_and_contexts_query(user, :all) do
+    from(t in UserToken, where: t.user_id == ^user.id)
+  end
+
+  def user_and_contexts_query(user, [_ | _] = contexts) do
+    from(t in UserToken,
+      where: t.user_id == ^user.id and t.context in ^contexts
+    )
+  end
+end
